Day 6 part 2 completed

+++ src/main/java/com/magicalpoet/advent/twentythree/Day06Races.java
@@ -3,19 +3,19 @@ package com.magicalpoet.advent.twentythree;
 import java.util.Map;
 
 public class Day06Races {
-    public int scoreRaces(Map<Integer, Integer> race) {
-        int totalWays = 1;
+    public long scoreRaces(Map<Long, Long> race) {
+        long totalWays = 1;
         for (var entry : race.entrySet()) {
             totalWays = totalWays * waysToWin(entry.getKey(), entry.getValue());
         }
         return totalWays;
     }
 
-    public int waysToWin(Integer time, Integer recordDistance) {
-        int holdTime = time / 2;
-        int waysToWin = 0;
+    public long waysToWin(Long time, Long recordDistance) {
+        long holdTime = time / 2;
+        long waysToWin = 0;
         while (holdTime > 0) {
-            int distanceTraveled = holdTime * (time - holdTime);
+            long distanceTraveled = holdTime * (time - holdTime);
             if (distanceTraveled > recordDistance) {
                 waysToWin++;
                 holdTime--;
